Using the new version of data bag search from https://github.com/edelight/chef-solo-search
