+++ MANIFEST.in
@@ -1,5 +1,6 @@
 include README.md
 include LICENSE
 include NOTICE
-include littlechef/data_bags_patch.rb
+include littlechef/data_bags.rb
+include littlechef/search.rb
 include littlechef/solo.rb
