+++ littlechef/search.rb
@@ -1,20 +1,39 @@
-#Copyright 2011 Markus Korn <thekorn@gmx.de>
 #
-#Licensed under the Apache License, Version 2.0 (the "License");
-#you may not use this file except in compliance with the License.
-#You may obtain a copy of the License at
+# Copyright 2011, edelight GmbH
 #
-#    http://www.apache.org/licenses/LICENSE-2.0
+# Authors:
+#       Markus Korn <markus.korn@edelight.de>
 #
-#Unless required by applicable law or agreed to in writing, software
-#distributed under the License is distributed on an "AS IS" BASIS,
-#WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#See the License for the specific language governing permissions and
-#limitations under the License.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# based on Brian Akins's patch: http://lists.opscode.com/sympa/arc/chef/2011-02/msg00000.html
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 
 if Chef::Config[:solo]
+  
+  if (defined? require_relative).nil?
+    # defenition of 'require_relative' for ruby < 1.9, found on stackoverflow.com
+    def require_relative(relative_feature)
+      c = caller.first
+      fail "Can't parse #{c}" unless c.rindex(/:\d+(:in `.*')?$/)
+      file = $`
+      if /\A\((.*)\)/ =~ file # eval, etc.
+        raise LoadError, "require_relative is called in #{$1}"
+      end
+      absolute = File.expand_path(relative_feature, File.dirname(file))
+      require absolute
+    end
+  end
+  
+  require_relative 'data_bags.rb'
   
   # Checks if a given `value` is equal to `match`
   # If value is an Array, then `match` is checked against each of the value's
@@ -171,35 +190,6 @@ if Chef::Config[:solo]
       end
     end
   end
-  
-  class Chef
-    module Mixin
-      module Language
-        # Hook into Chef which reads all items in a given `bag` and converts
-        # them into one single Hash
-        def data_bag(bag)
-          @solo_data_bags = {} if @solo_data_bags.nil?
-          unless @solo_data_bags[bag]
-            @solo_data_bags[bag] = {}
-            data_bag_path = Chef::Config[:data_bag_path]
-            Dir.glob(File.join(data_bag_path, bag, "*.json")).each do |f|
-              item = JSON.parse(IO.read(f))
-              @solo_data_bags[bag][item['id']] = item
-            end
-          end
-          @solo_data_bags[bag].keys
-        end
-
-        # Hook into Chef which returns the ruby representation of a given
-        # data_bag item
-        def data_bag_item(bag, item)
-          data_bag(bag) unless ( !@solo_data_bags.nil? && @solo_data_bags[bag])
-          @solo_data_bags[bag][item]
-        end
-
-      end
-    end
-  end
 
   class Chef
     class Recipe
@@ -243,4 +233,5 @@ if Chef::Config[:solo]
       end
     end
   end
+
 end
