+++ setup.py
@@ -21,7 +21,7 @@ setup(
     keywords=["chef", "devops"],
     install_requires=['fabric>=1.0.2', 'simplejson'],
     packages=['littlechef'],
-    package_data={'littlechef': ['data_bags_patch.rb', 'solo.rb']},
+    package_data={'littlechef': ['data_bags.rb', 'search.rb', 'solo.rb']},
     scripts=['cook'],
     test_suite='nose.collector',
     classifiers=[
